Fix Filter Key Track (3)

+++ DigitalSynthVRA8U/osc.h
@@ -398,10 +398,10 @@ public:
     if (m_mono_mode) {
       uint16_t shifted_pitch = (64 << 8) + m_pitch_current[0] + m_pitch_bend_normalized;
       uint16_t osc_pitch;
-      if (shifted_pitch > (64 << 8) + (NOTE_NUMBER_MAX << 8)) {
-        osc_pitch = (NOTE_NUMBER_MAX << 8);
-      } else if (shifted_pitch < (64 << 8) + (NOTE_NUMBER_MIN << 8)) {
-        osc_pitch = (NOTE_NUMBER_MIN << 8);
+      if (shifted_pitch > (64 << 8) + (static_cast<uint16_t>(NOTE_NUMBER_MAX) << 8)) {
+        osc_pitch = (static_cast<uint16_t>(NOTE_NUMBER_MAX) << 8);
+      } else if (shifted_pitch < (64 << 8) + (static_cast<uint16_t>(NOTE_NUMBER_MIN) << 8)) {
+        osc_pitch = (static_cast<uint16_t>(NOTE_NUMBER_MIN) << 8);
       } else {
         osc_pitch = m_pitch_current[0] + m_pitch_bend_normalized;
       }
